api/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9
bcrypt==4.0.1
python-dotenv